Revert types file move to make PR smaller
